% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/label_formatting.R
\name{abbreviate_word}
\alias{abbreviate_word}
\title{Abbreviate Word Using Rules}
\usage{
abbreviate_word(word, min_length = 8)
}
\arguments{
\item{word}{Character. Single word to abbreviate}

\item{min_length}{Integer. Minimum word length to abbreviate}
}
\value{
Character. Abbreviated word
}
\description{
Rule-based word abbreviation system.
}
\keyword{internal}
