% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/label_formatting.R
\name{apply_case_transform}
\alias{apply_case_transform}
\title{Apply Case Transformation}
\usage{
apply_case_transform(words, case)
}
\arguments{
\item{words}{Character vector. Words to transform}

\item{case}{Character. Case style to apply}
}
\value{
Character vector. Transformed words
}
\description{
Apply Case Transformation
}
\keyword{internal}
