% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gm_model_data.R
\name{build_analysis_data.Mclust}
\alias{build_analysis_data.Mclust}
\title{Build Analysis Data for Mclust Objects}
\usage{
\method{build_analysis_data}{Mclust}(fit_results, interpretation_args = NULL, ...)
}
\arguments{
\item{fit_results}{An object of class "Mclust" from the mclust package}

\item{interpretation_args}{List of interpretation parameters (optional)}

\item{...}{Additional arguments passed to interpretation_args_gm}
}
\value{
A list containing standardized GM data for interpretation
}
\description{
Extracts and standardizes data from fitted mclust Gaussian mixture models.
}
\keyword{internal}
