% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s3_model_data.R
\name{build_model_data.default}
\alias{build_model_data.default}
\title{Default method for build_model_data}
\usage{
\method{build_model_data}{default}(
  fit_results,
  model_type = NULL,
  interpretation_args = NULL,
  ...
)
}
\description{
Default method for build_model_data
}
\keyword{internal}
