% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gm_report.R
\name{build_model_info_section_gm}
\alias{build_model_info_section_gm}
\title{Build Model Information Section for GM Report}
\usage{
build_model_info_section_gm(analysis_data, fit_summary, format, heading_level)
}
\arguments{
\item{analysis_data}{Standardized GM analysis data}

\item{fit_summary}{Diagnostic information}

\item{format}{Output format}

\item{heading_level}{Heading level for markdown}
}
\value{
Character string with model information
}
\description{
Build Model Information Section for GM Report
}
\keyword{internal}
