% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generic_interpret.R
\name{build_report.default}
\alias{build_report.default}
\title{Default method for build_report}
\usage{
\method{build_report}{default}(
  interpretation,
  output_format = "cli",
  heading_level = 1,
  suppress_heading = FALSE,
  ...
)
}
\description{
Default method for build_report
}
\keyword{internal}
