% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fa_model_data.R
\name{build_structured_list.fa}
\alias{build_structured_list.fa}
\title{Build Structured List for FA}
\usage{
\method{build_structured_list}{fa}(x, analysis_type, ...)
}
\description{
Build Structured List for FA
}
\keyword{internal}
