% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gm_report.R
\name{build_title_section_gm}
\alias{build_title_section_gm}
\title{Build Title Section for GM Report}
\usage{
build_title_section_gm(analysis_data, format, heading_level)
}
\arguments{
\item{analysis_data}{Standardized GM analysis data}

\item{format}{Output format}

\item{heading_level}{Heading level for markdown}
}
\value{
Character string with title section
}
\description{
Build Title Section for GM Report
}
\keyword{internal}
