% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gm_prompt_builder.R
\name{build_variable_section_gm}
\alias{build_variable_section_gm}
\title{Build Variable Section for GM Prompt}
\usage{
build_variable_section_gm(analysis_data, variable_info, llm_args)
}
\arguments{
\item{analysis_data}{Standardized GM analysis data}

\item{variable_info}{Data frame with variable descriptions}

\item{llm_args}{List of LLM configuration parameters}
}
\value{
Character string with formatted variable information
}
\description{
Build Variable Section for GM Prompt
}
\keyword{internal}
