% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/label_json.R
\name{create_default_labels}
\alias{create_default_labels}
\title{Create Default Labels}
\usage{
create_default_labels(variable_info)
}
\arguments{
\item{variable_info}{Data frame. Variable information}
}
\value{
List. Default label structure
}
\description{
Create Default Labels
}
\keyword{internal}
