% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generic_interpret.R
\name{create_diagnostics.default}
\alias{create_diagnostics.default}
\title{Default method for create_diagnostics}
\usage{
\method{create_diagnostics}{default}(model_type, model_data, variable_info, ...)
}
\description{
Default method for create_diagnostics
}
\keyword{internal}
