% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/config.R
\name{default_fa_args}
\alias{default_fa_args}
\title{Get Default FA Config}
\usage{
default_fa_args()
}
\value{
fa_args object with default settings
}
\description{
Returns default factor analysis configuration. Useful for programmatic access.
}
