% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/label_utils.R
\name{detect_case_pattern}
\alias{detect_case_pattern}
\title{Detect Case Pattern of Word}
\usage{
detect_case_pattern(word)
}
\arguments{
\item{word}{Character. Word to analyze}
}
\value{
Character. Case pattern identifier
}
\description{
Detect Case Pattern of Word
}
\keyword{internal}
