% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s3_export.R
\name{export_interpretation.default}
\alias{export_interpretation.default}
\title{Default Export Method}
\usage{
\method{export_interpretation}{default}(
  interpretation_results,
  format = "txt",
  file = "interpretation",
  silent = 0
)
}
\description{
Default Export Method
}
\keyword{internal}
