% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s3_parameter_extraction.R
\name{extract_model_parameters.default}
\alias{extract_model_parameters.default}
\title{Default Method for extract_model_parameters}
\usage{
\method{extract_model_parameters}{default}(analysis_type, interpretation_args, ...)
}
\arguments{
\item{analysis_type}{Object with model type class}

\item{interpretation_args}{Config object}

\item{...}{Additional arguments}
}
\value{
Empty list
}
\description{
Returns an empty list for unknown model types.
}
\keyword{internal}
