% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gm_visualization.R
\name{extract_variance_matrix}
\alias{extract_variance_matrix}
\title{Extract Variance Matrix from GM Analysis Data}
\usage{
extract_variance_matrix(analysis_data)
}
\arguments{
\item{analysis_data}{Standardized GM analysis data containing covariances}
}
\value{
Matrix of standard deviations (n_variables x n_clusters)
}
\description{
Extracts within-cluster standard deviations from covariance matrices.
}
\keyword{internal}
