% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gm_visualization.R
\name{extract_variance_ratio_matrix}
\alias{extract_variance_ratio_matrix}
\title{Extract Variance Ratio Matrix from GM Analysis Data}
\usage{
extract_variance_ratio_matrix(analysis_data)
}
\arguments{
\item{analysis_data}{Standardized GM analysis data containing means and covariances}
}
\value{
Matrix of variance ratios (n_variables x n_clusters)
}
\description{
Calculates between-cluster to within-cluster variance ratios for each variable.
Higher ratios indicate variables that better discriminate between clusters.
}
\keyword{internal}
