% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gm_visualization.R
\name{filter_variables_gm}
\alias{filter_variables_gm}
\title{Filter Variables for GM Visualization}
\usage{
filter_variables_gm(analysis_data, var_indices)
}
\arguments{
\item{analysis_data}{Standardized GM analysis data}

\item{var_indices}{Indices of variables to keep}
}
\value{
Modified analysis_data with filtered variables
}
\description{
Helper function to filter analysis_data to specific variables.
}
\keyword{internal}
