% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gm_diagnostics.R
\name{find_distinguishing_variables_gm}
\alias{find_distinguishing_variables_gm}
\title{Find Distinguishing Variables for Clusters}
\usage{
find_distinguishing_variables_gm(analysis_data, top_n = 5)
}
\arguments{
\item{analysis_data}{Standardized GM analysis data}

\item{top_n}{Number of top distinguishing variables per cluster}
}
\value{
List with distinguishing variables for each cluster
}
\description{
Identifies variables that best distinguish each cluster from others.
}
\keyword{internal}
