% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core_parameter_registry.R
\name{get_param_default}
\alias{get_param_default}
\title{Get Parameter Default Value}
\usage{
get_param_default(param_name)
}
\arguments{
\item{param_name}{Character. Name of the parameter}
}
\value{
Default value for the parameter
}
\description{
Retrieves the default value for a parameter from the registry.
}
\examples{
get_param_default("word_limit")
get_param_default("cutoff")
}
