% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interpret_method_dispatch.R
\name{interpret_model}
\alias{interpret_model}
\title{Internal S3 Generic for Interpreting Fitted Model Objects}
\usage{
interpret_model(model, variable_info, ...)
}
\arguments{
\item{model}{Fitted model object}

\item{variable_info}{Variable descriptions dataframe}

\item{...}{Additional arguments}
}
\value{
Interpretation object
}
\description{
This is an internal generic used by interpret() to dispatch on fitted
model objects from various packages. Not exported - users should call
interpret() directly.
}
\keyword{internal}
