% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chat_fa.R
\name{is.chat_fa}
\alias{is.chat_fa}
\title{Check if object is a chat_fa}
\usage{
is.chat_fa(x)
}
\arguments{
\item{x}{Object to test}
}
\value{
Logical indicating if x is a chat_fa object
}
\description{
Check if object is a chat_fa
}
