% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_interpretation.R
\name{is.interpretation}
\alias{is.interpretation}
\title{Check if object is an interpretation}
\usage{
is.interpretation(x)
}
\arguments{
\item{x}{Object to test}
}
\value{
Logical indicating if x is an interpretation object
}
\description{
Check if object is an interpretation
}
