% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/label_json.R
\name{is_nonempty_string}
\alias{is_nonempty_string}
\title{Is x a Single Non-Blank Character Value?}
\usage{
is_nonempty_string(x)
}
\arguments{
\item{x}{Value to test}
}
\value{
Logical scalar
}
\description{
Is x a Single Non-Blank Character Value?
}
\keyword{internal}
