% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/label_json.R
\name{order_labels}
\alias{order_labels}
\title{Reorder Label Records to Match variable_info}
\usage{
order_labels(parsed, variable_info)
}
\arguments{
\item{parsed}{List of validated label records}

\item{variable_info}{Data frame. Original variables}
}
\value{
List of records in \code{variable_info$variable} order
}
\description{
Reorder Label Records to Match variable_info
}
\keyword{internal}
