% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chat_fa.R
\name{print.chat_fa}
\alias{print.chat_fa}
\title{Print method for chat_fa objects}
\usage{
\method{print}{chat_fa}(x, ...)
}
\arguments{
\item{x}{A chat_fa object}

\item{...}{Additional arguments (unused)}
}
\description{
Print method for chat_fa objects
}
