% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/base_chat_session.R
\name{print.chat_session}
\alias{print.chat_session}
\title{Print method for chat_session objects}
\usage{
\method{print}{chat_session}(x, ...)
}
\arguments{
\item{x}{A chat_session object}

\item{...}{Additional arguments (unused)}
}
\description{
Print method for chat_session objects
}
