% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shared_config.R
\name{print.interpretation_args}
\alias{print.interpretation_args}
\title{Print method for interpretation_args}
\usage{
\method{print}{interpretation_args}(x, ...)
}
\description{
Print method for interpretation_args
}
\keyword{internal}
