% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/label_args.R
\name{print.label_args}
\alias{print.label_args}
\title{Print Label Arguments}
\usage{
\method{print}{label_args}(x, ...)
}
\arguments{
\item{x}{A label_args object}

\item{...}{Additional arguments (unused)}
}
\description{
Print Label Arguments
}
