% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/labeling_args.R
\name{print.labeling_args}
\alias{print.labeling_args}
\title{Print Labeling Arguments}
\usage{
\method{print}{labeling_args}(x, ...)
}
\arguments{
\item{x}{A labeling_args object}

\item{...}{Additional arguments (unused)}
}
\description{
Print Labeling Arguments
}
