% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shared_print.R
\name{print_header}
\alias{print_header}
\title{Create compact header for config/session objects}
\usage{
print_header(title, width = .PRINT_WIDTH_COMPACT)
}
\arguments{
\item{title}{Header title text}

\item{width}{Width of the underline rule}
}
\description{
Create compact header for config/session objects
}
\keyword{internal}
