% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shared_print.R
\name{print_item}
\alias{print_item}
\title{Create bullet item}
\usage{
print_item(text, indent = .PRINT_INDENT)
}
\arguments{
\item{text}{Item text}

\item{indent}{Number of spaces to indent}
}
\description{
Create bullet item
}
\keyword{internal}
