% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shared_print.R
\name{print_kv}
\alias{print_kv}
\title{Create key-value line}
\usage{
print_kv(key, value, indent = .PRINT_INDENT)
}
\arguments{
\item{key}{The label/key}

\item{value}{The value to display}

\item{indent}{Number of spaces to indent}
}
\description{
Create key-value line
}
\keyword{internal}
