% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shared_print.R
\name{print_section}
\alias{print_section}
\title{Create section header (no underline)}
\usage{
print_section(title, indent = 0)
}
\arguments{
\item{title}{Section title}

\item{indent}{Number of spaces to indent}
}
\description{
Create section header (no underline)
}
\keyword{internal}
