% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/label_formatting.R
\name{remove_label_articles}
\alias{remove_label_articles}
\title{Remove Articles from Label}
\usage{
remove_label_articles(text)
}
\arguments{
\item{text}{Character. Label text}
}
\value{
Character. Text without articles
}
\description{
Remove Articles from Label
}
\keyword{internal}
