% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/label_formatting.R
\name{remove_label_prepositions}
\alias{remove_label_prepositions}
\title{Remove Prepositions from Label}
\usage{
remove_label_prepositions(text)
}
\arguments{
\item{text}{Character. Label text}
}
\value{
Character. Text without prepositions
}
\description{
Remove Prepositions from Label
}
\keyword{internal}
