% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/label_utils.R
\name{restore_case_pattern}
\alias{restore_case_pattern}
\title{Restore Case Pattern to Word}
\usage{
restore_case_pattern(word, pattern)
}
\arguments{
\item{word}{Character. Word to transform}

\item{pattern}{Character. Case pattern to apply}
}
\value{
Character. Word with case pattern applied
}
\description{
Restore Case Pattern to Word
}
\keyword{internal}
