% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/label_json.R
\name{simplify_description}
\alias{simplify_description}
\title{Simplify Description to Label}
\usage{
simplify_description(description)
}
\arguments{
\item{description}{Character. Variable description}
}
\value{
Character. Simplified label
}
\description{
Basic heuristic to create a label from a description.
}
\keyword{internal}
