% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gm_report.R
\name{style_for_cli_gm}
\alias{style_for_cli_gm}
\title{Style Report for CLI Output (GM)}
\usage{
style_for_cli_gm(report)
}
\arguments{
\item{report}{Character string with report content}
}
\value{
Styled report for CLI display
}
\description{
Style Report for CLI Output (GM)
}
\keyword{internal}
