% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s3_list_validation.R
\name{validate_list_structure.default}
\alias{validate_list_structure.default}
\title{Default Method for validate_list_structure}
\usage{
\method{validate_list_structure}{default}(model_type, fit_results_list)
}
\arguments{
\item{model_type}{Dispatch object with model type class}

\item{fit_results_list}{List. Structured list}
}
\value{
Never returns - always errors
}
\description{
Returns an error for unsupported model types.
}
\keyword{internal}
