% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s3_parameter_extraction.R
\name{validate_model_requirements.default}
\alias{validate_model_requirements.default}
\title{Default Method for validate_model_requirements}
\usage{
\method{validate_model_requirements}{default}(analysis_type, ...)
}
\arguments{
\item{analysis_type}{Object with model type class}

\item{...}{Additional arguments}
}
\value{
Invisible NULL
}
\description{
No validation for unknown model types.
}
\keyword{internal}
