% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s3_parameter_extraction.R
\name{validate_model_requirements.gm}
\alias{validate_model_requirements.gm}
\title{Validate GM Requirements (Placeholder)}
\usage{
\method{validate_model_requirements}{gm}(analysis_type, ...)
}
\arguments{
\item{analysis_type}{GM analysis type object}

\item{...}{Additional arguments}
}
\value{
Invisible NULL
}
\description{
Validate GM Requirements (Placeholder)
}
\keyword{internal}
