% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s3_prompt_builder.R
\name{build_main_prompt.default}
\alias{build_main_prompt.default}
\title{Default method for build_main_prompt}
\usage{
\method{build_main_prompt}{default}(model_type, model_data, variable_info, ...)
}
\description{
Default method for build_main_prompt
}
\keyword{internal}
