% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generic_interpret.R
\name{build_report}
\alias{build_report}
\title{Build Report (S3 Generic)}
\usage{
build_report(
  interpretation,
  output_format = "text",
  heading_level = 1,
  suppress_heading = FALSE,
  ...
)
}
\arguments{
\item{interpretation}{Interpretation object}

\item{output_format}{Character. "text" or "markdown"}

\item{heading_level}{Integer. Markdown heading level}

\item{suppress_heading}{Logical. Suppress report heading}

\item{...}{Additional arguments for model-specific report building}
}
\value{
Character. Formatted report
}
\description{
Model-specific report building
}
\keyword{internal}
