% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generic_prompt_builder.R
\name{build_system_prompt.default}
\alias{build_system_prompt.default}
\title{Default method for build_system_prompt}
\usage{
\method{build_system_prompt}{default}(model_type, word_limit, ...)
}
\description{
Default method for build_system_prompt
}
\keyword{internal}
