% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gm_diagnostics.R
\name{calculate_cluster_separation_gm}
\alias{calculate_cluster_separation_gm}
\title{Calculate Cluster Separation for GM}
\usage{
calculate_cluster_separation_gm(analysis_data)
}
\arguments{
\item{analysis_data}{Standardized GM analysis data}
}
\value{
Matrix of pairwise distances or NULL if not computable
}
\description{
Computes pairwise Mahalanobis distances between cluster centers.
}
\keyword{internal}
