% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s3_json_parser.R
\name{create_default_result.default}
\alias{create_default_result.default}
\title{Default method for create_default_result}
\usage{
\method{create_default_result}{default}(analysis_type, ...)
}
\description{
Default method for create_default_result
}
\keyword{internal}
