% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/base_chat_session.R
\name{is.chat_fa}
\alias{is.chat_fa}
\title{Backward compatibility: is.chat_fa}
\usage{
is.chat_fa(x)
}
\arguments{
\item{x}{Object to test}
}
\value{
Logical
}
\description{
Backward compatibility: is.chat_fa
}
