% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/base_chat_session.R
\name{is.chat_session}
\alias{is.chat_session}
\title{Check if object is a chat_session}
\usage{
is.chat_session(x)
}
\arguments{
\item{x}{Object to test}
}
\value{
Logical indicating if x is a chat_session object
}
\description{
Check if object is a chat_session
}
