% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/label_formatting.R
\name{print.variable_labels}
\alias{print.variable_labels}
\title{Print Variable Labels Object}
\usage{
\method{print}{variable_labels}(x, silent = 0, ...)
}
\arguments{
\item{x}{variable_labels object}

\item{silent}{Integer. Controls output verbosity (0, 1, 2). Default = 0}

\item{...}{Additional arguments}
}
\description{
Print Variable Labels Object
}
