% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_chat_session.R
\name{reset.chat_fa}
\alias{reset.chat_fa}
\title{Backward compatibility: reset.chat_fa}
\usage{
reset.chat_fa(chat_obj)
}
\arguments{
\item{chat_obj}{A chat_fa object}
}
\value{
Reset chat_session
}
\description{
Backward compatibility: reset.chat_fa
}
