% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s3_json_parser.R
\name{validate_parsed_result.default}
\alias{validate_parsed_result.default}
\title{Default method for validate_parsed_result}
\usage{
\method{validate_parsed_result}{default}(parsed, analysis_type, ...)
}
\description{
Default method for validate_parsed_result
}
\keyword{internal}
